/proc/stat, partial meminfo and cleanups
This adds support for /proc/stat which combined with the existing
/proc/cpuinfo support now gives the expected output in a way that's
properly parsed by top and htop.

Basic meminfo support is also added, with support for MemTotal and
MemFree, a potentially slightly off MemAvailable, Cached pages and
return 0 bytes for a few fields we don't have data for.

Also add a bunch of commands and add the start option to have it
readable by other users (required for unprivileged containers).

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>